chore(BallotBoxAnimated): moved component to layout
