+++ src/components/Layout/BallotBoxAnimated.test.tsx
