+++ src/components/Layout/BallotBoxAnimated.tsx
