+++ src/components/Process/View.tsx
@@ -30,9 +30,9 @@ import { Trans, useTranslation } from 'react-i18next'
 import { RiErrorWarningLine } from 'react-icons/ri'
 import { FacebookShare, RedditShare, TelegramShare, TwitterShare } from '~components/Share'
 import { ElectionQuestions, ElectionResults, environment } from '~components/vocdoni-ui'
+import { BallotBoxAnimated } from '../Layout/BallotBoxAnimated'
 import { ActionsMenu } from './ActionsMenu'
 import ProcessAside, { VoteButton } from './Aside'
-import { BallotBoxAnimated } from './BallotBoxAnimated'
 import { ConfirmVoteModal } from './ConfirmVoteModal'
 import { CreatedBy } from './CreatedBy'
 import { ElectionVideo } from './Dashboard/ProcessView'
